backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
requests==2.31.0
pillow==10.2.0
pytest==8.0.0
pytest-cov==4.1.0 